backend/requirements.txt
pillow>=9.0.0
numpy==1.26.2
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.29.0
python-multipart==0.0.6
